refactor: move organization service and organization-user response to admin-console, refs AC-1202
