+++ apps/browser/src/admin-console/background/service-factories/organization-service.factory.ts
@@ -9,7 +9,7 @@ import {
   stateServiceFactory,
   StateServiceInitOptions,
 } from "../../../background/service_factories/state-service.factory";
-import { BrowserOrganizationService } from "../../../services/browser-organization.service";
+import { BrowserOrganizationService } from "../../services/browser-organization.service";
 
 type OrganizationServiceFactoryOptions = FactoryOptions;
