+++ apps/browser/src/admin-console/services/browser-organization.service.ts
@@ -3,7 +3,7 @@ import { BehaviorSubject } from "rxjs";
 import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
 import { OrganizationService } from "@bitwarden/common/admin-console/services/organization/organization.service";
 
-import { browserSession, sessionSync } from "../decorators/session-sync-observable";
+import { browserSession, sessionSync } from "../../decorators/session-sync-observable";
 
 @browserSession
 export class BrowserOrganizationService extends OrganizationService {
