+++ apps/cli/src/commands/list.command.ts
@@ -16,10 +16,10 @@ import { FolderService } from "@bitwarden/common/vault/abstractions/folder/folde
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 
 import { CollectionResponse } from "../admin-console/models/response/collection.response";
+import { OrganizationUserResponse } from "../admin-console/models/response/organization-user.response";
 import { OrganizationResponse } from "../admin-console/models/response/organization.response";
 import { Response } from "../models/response";
 import { ListResponse } from "../models/response/list.response";
-import { OrganizationUserResponse } from "../models/response/organization-user.response";
 import { CliUtils } from "../utils";
 import { CipherResponse } from "../vault/models/cipher.response";
 import { FolderResponse } from "../vault/models/folder.response";
